@@ -1,16 +1,22 @@
 module github.com/corazawaf/coraza/v3
 
-go 1.16
+go 1.18
 
 require (
 	github.com/anuraaga/go-modsecurity v0.0.0-20220816070944-f36055ce7d5d
 	github.com/corazawaf/libinjection-go v0.0.0-20220207031228-44e9c4250eb5
 	github.com/foxcpp/go-mockdns v1.0.0
 	github.com/magefile/mage v1.13.0
-	github.com/miekg/dns v1.1.50 // indirect
 	github.com/petar-dambovaliev/aho-corasick v0.0.0-20211021192214-5ab2d9280aa9
 	github.com/tidwall/gjson v1.14.2
 	golang.org/x/net v0.0.0-20220809184613-07c6da5e1ced
+)
+
+require (
+	github.com/miekg/dns v1.1.50 // indirect
+	github.com/tidwall/match v1.1.1 // indirect
+	github.com/tidwall/pretty v1.2.0 // indirect
+	golang.org/x/mod v0.6.0-dev.0.20220419223038-86c51ed26bb4 // indirect
 	golang.org/x/sys v0.0.0-20220808155132-1c4a2a72c664 // indirect
 	golang.org/x/tools v0.1.12 // indirect
 )
