@@ -5,7 +5,6 @@ package io
 
 import (
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path"
 )
@@ -18,13 +17,13 @@ func ReadFirstFile(directories []string, filename string) ([]byte, error) {
 	}
 	if filename[0] == '/' {
 		// filename is absolute
-		return ioutil.ReadFile(filename)
+		return os.ReadFile(filename)
 	}
 	for _, p := range directories {
 		f := path.Join(p, filename)
 		// if the file does exist we return it
 		if _, err := os.Stat(f); err == nil {
-			return ioutil.ReadFile(f)
+			return os.ReadFile(f)
 		}
 	}
 	return nil, fmt.Errorf("file %s not found", filename)
