@@ -5,7 +5,6 @@ package operators
 
 import (
 	"context"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"strconv"
@@ -33,7 +32,7 @@ func TestOperators(t *testing.T) {
 	}
 	if err := filepath.Walk(root, func(path string, info os.FileInfo, err error) error {
 		if strings.HasSuffix(path, ".json") {
-			data, _ := ioutil.ReadFile(path)
+			data, _ := os.ReadFile(path)
 			files = append(files, data)
 		}
 		return nil
