@@ -4,7 +4,6 @@
 package transformations
 
 import (
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"strconv"
@@ -31,7 +30,7 @@ func TestTransformations(t *testing.T) {
 	}
 	if err := filepath.Walk(root, func(path string, info os.FileInfo, err error) error {
 		if strings.HasSuffix(path, ".json") {
-			data, _ := ioutil.ReadFile(path)
+			data, _ := os.ReadFile(path)
 			files = append(files, data)
 		}
 		return nil
